[LeetCode Sync] Runtime - 0 ms (100.00%), Memory - 2.1 MB (89.77%)

+++ problems/4321-digit-frequency-score/solution.rs
@@ -13,7 +13,7 @@ impl Solution {
 
         freq
             .into_iter()
-            .map(|(digit, count)| (digit as i32) * (count as i32))
+            .map(|(digit, count)| digit * (count as i32))
             .sum()
     }
 }
